requirements.txt
google-adk==1.14.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
greenlet>=3.0.0
python-multipart>=0.0.9
